feat: invite role class and some clean ups
