+++ packages/discord.js/src/index.js
@@ -127,6 +127,7 @@ exports.BaseGuildTextChannel = require('./structures/BaseGuildTextChannel.js').B
 exports.BaseGuildVoiceChannel = require('./structures/BaseGuildVoiceChannel.js').BaseGuildVoiceChannel;
 exports.BaseInteraction = require('./structures/BaseInteraction.js').BaseInteraction;
 exports.BaseInvite = require('./structures/BaseInvite.js').BaseInvite;
+exports.BaseRole = require('./structures/BaseRole.js').BaseRole;
 exports.BaseSelectMenuComponent = require('./structures/BaseSelectMenuComponent.js').BaseSelectMenuComponent;
 exports.ButtonComponent = require('./structures/ButtonComponent.js').ButtonComponent;
 exports.ButtonInteraction = require('./structures/ButtonInteraction.js').ButtonInteraction;
@@ -181,6 +182,7 @@ exports.InteractionCallbackResponse =
 exports.InteractionCollector = require('./structures/InteractionCollector.js').InteractionCollector;
 exports.InteractionWebhook = require('./structures/InteractionWebhook.js').InteractionWebhook;
 exports.InviteGuild = require('./structures/InviteGuild.js').InviteGuild;
+exports.InviteRole = require('./structures/InviteRole.js').InviteRole;
 exports.LabelComponent = require('./structures/LabelComponent.js').LabelComponent;
 exports.MediaChannel = require('./structures/MediaChannel.js').MediaChannel;
 exports.MediaGalleryComponent = require('./structures/MediaGalleryComponent.js').MediaGalleryComponent;
