+++ packages/discord.js/src/structures/InviteRole.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const { BaseRole } = require('./BaseRole');
+
+/**
+ * Represents a role received from an invite
+ *
+ * @extends {BaseRole}
+ */
+class InviteRole extends BaseRole {}
+
+exports.InviteRole = InviteRole;
