+++ packages/discord.js/src/util/DataResolver.js
@@ -11,6 +11,30 @@ const { BaseInvite } = require('../structures/BaseInvite.js');
 // Fixes circular dependencies.
 const getGuildTemplate = lazy(() => require('../structures/GuildTemplate.js').GuildTemplate);
 
+// Ideally this should be placed in `Invite.js` util file, but causes circular dependency there.
+/**
+ * Creates form data body payload for invite
+ *
+ * @param {Client} client The client
+ * @param {InviteCreateOptions} options The options for creating invite
+ * @returns {Promise<FormData>}
+ * @ignore
+ */
+async function createInviteFormData(client, { targetUsersFile, ...rest } = {}) {
+  const formData = new FormData();
+  let usersCsv;
+  if (Array.isArray(targetUsersFile)) {
+    usersCsv = targetUsersFile.map(user => client.users.resolveId(user)).join('\n');
+  } else {
+    const resolved = await resolveFile(targetUsersFile);
+    usersCsv = resolved.data.toString('utf8');
+  }
+
+  formData.append('target_users_file', new Blob([usersCsv], { type: 'text/csv' }), 'users.csv');
+  formData.append('payload_json', JSON.stringify(rest));
+  return formData;
+}
+
 /**
  * Data that can be resolved to give an invite code. This can be:
  * - An invite code
@@ -159,3 +183,4 @@ exports.resolveGuildTemplateCode = resolveGuildTemplateCode;
 exports.resolveImage = resolveImage;
 exports.resolveBase64 = resolveBase64;
 exports.resolveFile = resolveFile;
+exports.createInviteFormData = createInviteFormData;
