+++ packages/discord.js/src/util/Transformers.js
@@ -116,9 +116,28 @@ function _transformCollectibles(collectibles) {
   };
 }
 
+/**
+ * Transforms API invite target users job status
+ *
+ * @param {RESTGetAPIInviteTargetUsersJobStatusResult} job The job status to transform
+ * @returns {TargetUsersJobStatusForInvite}
+ * @ignore
+ */
+function _transformAPIInviteTargetUsersJobStatus(job) {
+  return {
+    status: job.status,
+    totalUsers: job.total_users,
+    processedUsers: job.processed_users,
+    createdAt: job.created_at ? new Date(job.created_at) : null,
+    completedAt: job.completed_at ? new Date(job.completed_at) : null,
+    errorMessage: job.error_message ?? null,
+  };
+}
+
 exports.toSnakeCase = toSnakeCase;
 exports._transformAPIAutoModerationAction = _transformAPIAutoModerationAction;
 exports._transformAPIMessageInteractionMetadata = _transformAPIMessageInteractionMetadata;
 exports._transformGuildScheduledEventRecurrenceRule = _transformGuildScheduledEventRecurrenceRule;
 exports._transformAPIIncidentsData = _transformAPIIncidentsData;
 exports._transformCollectibles = _transformCollectibles;
+exports._transformAPIInviteTargetUsersJobStatus = _transformAPIInviteTargetUsersJobStatus;
